+++ govcd/nsxt_manager.go
@@ -5,6 +5,9 @@
 package govcd
 
 import (
+	"fmt"
+	"net/http"
+
 	"github.com/vmware/go-vcloud-director/v2/types/v56"
 )
 
@@ -16,4 +19,48 @@ type NsxtManager struct {
 	//
 	// Note:  this is being computed when retrieving the structure and will not be populated if this structure is initialized manually
 	Urn string
+}
+
+// GetNsxtManagerByName searches for NSX-T managers available in VCD and returns the one that
+// matches name
+func (vcdClient *VCDClient) GetNsxtManagerByName(name string) (*NsxtManager, error) {
+	nsxtManagers, err := vcdClient.QueryNsxtManagerByName(name)
+	if err != nil {
+		return nil, fmt.Errorf("error retrieving NSX-T Manager by name '%s': %s", name, err)
+	}
+
+	// Double check that exactly one NSX-T Manager is found and throw error otherwise
+	singleNsxtManager, err := oneOrError("name", name, nsxtManagers)
+	if err != nil {
+		return nil, err
+	}
+
+	resp, err := vcdClient.Client.executeJsonRequest(singleNsxtManager.HREF, http.MethodGet, nil, "error retrieving NSX-T Manager: %s")
+	if err != nil {
+		return nil, err
+	}
+
+	defer closeBody(resp)
+
+	nsxtManager := NsxtManager{
+		NsxtManager: &types.NsxtManager{},
+		VCDClient:   vcdClient,
+	}
+
+	err = decodeBody(types.BodyTypeJSON, resp, nsxtManager.NsxtManager)
+	if err != nil {
+		return nil, err
+	}
+
+	// Populate computed URN field, which is not usually provided in API
+	nsxtManager.Urn = nsxtManager.NsxtManager.ID
+	if !isUrn(nsxtManager.NsxtManager.ID) {
+		nsxtManagerUrn, err := BuildUrnWithUuid("urn:vcloud:nsxtmanager:", nsxtManager.NsxtManager.ID)
+		if err != nil {
+			return nil, fmt.Errorf("error building NSX-T Manager URN from ID '%s': %s", nsxtManager.NsxtManager.ID, err)
+		}
+		nsxtManager.Urn = nsxtManagerUrn
+	}
+
+	return &nsxtManager, nil
 }
