+++ govcd/nsxt_manager_test.go
@@ -8,6 +8,7 @@ import (
 
 func (vcd *TestVCD) Test_NsxtManager(check *C) {
 	skipNoNsxtConfiguration(vcd, check)
+	vcd.skipIfNotSysAdmin(check)
 
 	nsxtManager, err := vcd.client.GetNsxtManagerByName(vcd.config.VCD.Nsxt.Manager)
 	check.Assert(err, IsNil)
