+++ govcd/system.go
@@ -939,49 +939,6 @@ func (vcdClient *VCDClient) QueryNsxtManagerByName(name string) ([]*types.QueryR
 	return results.Results.NsxtManagerRecord, nil
 }
 
-// QueryNsxtManagerByName searches for NSX-T managers available in VCD
-func (vcdClient *VCDClient) GetNsxtManagerByName(name string) (*NsxtManager, error) {
-	nsxtManagers, err := vcdClient.QueryNsxtManagerByName(name)
-	if err != nil {
-		return nil, fmt.Errorf("error retrieving NSX-T Manager by name '%s': %s", name, err)
-	}
-
-	// Double check that exactly one NSX-T Manager is found and throw error otherwise
-	singleNsxtManager, err := oneOrError("name", name, nsxtManagers)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := vcdClient.Client.executeJsonRequest(singleNsxtManager.HREF, http.MethodGet, nil, "error retrieving NSX-T Manager: %s")
-	if err != nil {
-		return nil, err
-	}
-
-	defer closeBody(resp)
-
-	nsxtManager := NsxtManager{
-		NsxtManager: &types.NsxtManager{},
-		VCDClient:   vcdClient,
-	}
-
-	err = decodeBody(types.BodyTypeJSON, resp, nsxtManager.NsxtManager)
-	if err != nil {
-		return nil, err
-	}
-
-	// Populate computed URN field, which is not usually provided in API
-	nsxtManager.Urn = nsxtManager.NsxtManager.ID
-	if !isUrn(nsxtManager.NsxtManager.ID) {
-		nsxtManagerUrn, err := BuildUrnWithUuid("urn:vcloud:nsxtmanager:", nsxtManager.NsxtManager.ID)
-		if err != nil {
-			return nil, fmt.Errorf("error building NSX-T Manager URN from ID '%s': %s", nsxtManager.NsxtManager.ID, err)
-		}
-		nsxtManager.Urn = nsxtManagerUrn
-	}
-
-	return &nsxtManager, nil
-}
-
 // QueryNsxtManagerByHref searches for NSX-T managers available in VCD
 func (vcdClient *VCDClient) QueryNsxtManagerByHref(href string) ([]*types.QueryResultNsxtManagerRecordType, error) {
 	results, err := vcdClient.QueryWithNotEncodedParams(nil, map[string]string{
